pyproject.toml
[tool.poetry]
name = "rng_bot"
version = "0.1.0"
description = ""
authors = ["syongkheng <33014950+syongkheng@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
python-telegram-bot = "20.0a2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
